# asklepios-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Asklepios ERP API Documentation",
        "description": "Documentation interactive de l'API Asklepios pour la gestion hospitalière. Permet aux développeurs front-end de tester les requêtes en temps réel.",
        "contact": {
            "name": "Support Technique Asklepios",
            "email": "admin@asklepios.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur API de Développement"
        }
    ],
    "paths": {
...
        "/api/admin/pharmaciens": {
            "get": {
                "tags": [
                    "Pharmaciens (Admin)"
                ],
                "summary": "Lister les pharmaciens",
                "description": "Lister et filtrer les pharmaciens (Sans pagination)",
                "operationId": "getAdminPharmaciens",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Nom, email ou téléphone",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "position",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "magasin",
                                "vente"
                            ]
                        }
                    },
                    {
                        "name": "branch_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste récupérée avec succès"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Pharmaciens (Admin)"
                ],
                "summary": "Créer un pharmacien",
                "description": "Créer un pharmacien (Compte User + Profil)",
                "operationId": "storeAdminPharmacien",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "phone",
                                    "email",
                                    "password",
                                    "position",
                                    "branch_id"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "phone": {
                                        "type": "integer"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "position": {
                                        "type": "string",
                                        "enum": [
                                            "magasin",
                                            "vente"
                                        ]
                                    },
                                    "branch_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pharmacien créé avec succès"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/pharmaciens/paginated": {
            "get": {
                "tags": [
                    "Pharmaciens (Admin)"
                ],
                "summary": "Lister les pharmaciens avec pagination",
                "description": "Lister et filtrer les pharmaciens (AVEC pagination)",
                "operationId": "getAdminPharmaciensPaginated",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Nom, email ou téléphone",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "position",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "magasin",
                                "vente"
                            ]
                        }
                    },
                    {
                        "name": "branch_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Nombre d'éléments par page (défaut: 15)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste paginée récupérée avec succès"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/pharmaciens/{id}": {
            "put": {
                "tags": [
                    "Pharmaciens (Admin)"
                ],
                "summary": "Modifier un pharmacien",
                "description": "Modifier un pharmacien",
                "operationId": "updateAdminPharmacien",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "phone": {
                                        "type": "integer"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "position": {
                                        "type": "string",
                                        "enum": [
                                            "magasin",
                                            "vente"
                                        ]
                                    },
                                    "branch_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pharmacien modifié avec succès"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Pharmaciens (Admin)"
                ],
                "summary": "Supprimer un pharmacien",
                "description": "Supprimer un pharmacien",
                "operationId": "deleteAdminPharmacien",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pharmacien supprimé avec succès"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Saisissez le jeton (Token) obtenu lors de la connexion pour tester les routes protégées.",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}